refactor: 용어 조회수 랭킹 DTO 패키지 변경
