+++ space-d/src/main/java/com/dnd/spaced/core/word/domain/dto/ViewCountStatisticsRankDto.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.word.domain.repository.dto.response;
+package com.dnd.spaced.core.word.domain.dto;
 
 public record ViewCountStatisticsRankDto(int rank, Long wordId, long viewCount) {
 }
