+++ space-d/src/main/java/com/dnd/spaced/core/word/domain/repository/WordViewCountStatisticsRepository.java
@@ -1,7 +1,7 @@
 package com.dnd.spaced.core.word.domain.repository;
 
 import com.dnd.spaced.core.word.domain.repository.dto.WordViewCountStatisticsDto;
-import com.dnd.spaced.core.word.domain.repository.dto.response.ViewCountStatisticsRankDto;
+import com.dnd.spaced.core.word.domain.dto.ViewCountStatisticsRankDto;
 import java.time.LocalDateTime;
 import java.util.List;
